+++ CHANGELOG.md
@@ -15,7 +15,7 @@
 
 ## Version 1.0.0
 *2017-07-14*
-* Remove guava and android plugin dependency.
+* Remove guava and android plugin variant.
 * Fix: when multiple proxy appear in one method, only one works.
 
 ## Version 1.0.0-beta1